+++ scribo_ct/examples/example_1/source/quux.c
@@ -3,13 +3,14 @@
 // File: quux.c
 // ---------------------------------------------------------------------------------------------------------------------
 
-// Produce log messages with automatically-filled message text
+// Produce log messages with default verbosity and/or automatically-filled message text
 
 #define SCRIBO_CATEGORY QUUX
 #include <scribo.h>
 
 void quux()
 {
+    SCRIBOD("Auto-fill method and trace:");
     SCRIBOM();
     SCRIBOT();
 }
@@ -18,6 +19,10 @@ void doQuux()
 {
     SCRIBO(LOG, "Quux...");
     
+    SCRIBO(DEBUG, "Auto-fill verbosity:");
+    SCRIBO();
+    
+    SCRIBO(DEBUG, "Auto-fill method and trace:");
     SCRIBO(METHOD);
     SCRIBO(TRACE);