+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# v?.?.?
+?/?/?
+- Made verbosity optional, defaulted to TRACE
+
 # v0.6
 2015/9/27
 - Added configuration option to suppress auto-fill of log message text